refactor: DocumentPage 내부의 컴포넌트 생성 로직 분리
